Determine C++ Filesystem implementation at CMake time
GitCondDB's build assumes GCC as the compiler, including its support
levels for C++ Filesystem in libstdc++. Builds on macOS with Xcode/Clang
thus fail due to direct use of lidstdc++fs and possible incorrect
inclusion of the appropriate filesystem header.

Use try_compile to determine if the compiler/standard library supports
std::filesystem directly or through a support/experimental library.
Only GNU's stdc++fs and LLVM's libc++fs are checked for when native
support is not available. A fallback to using Boost's filesystem library
is used otherwise.

A compile definition is used to select Boost or native support.
Currently this must be passed to tests due it being private to GitCondDB
but the tests exercise the internals of the library.

Implement calls to filesystem constructors/member functions/operators that
use string_view based on use or not of Boost due to lack of string_view
versions of these.

Include an optional check in GitCondDBConfig to refind Boost and the
needed Boost::filesystem target if required.

TODO: Currently tested on macOS High Sierra, Xcode 10, Boost 1.69.
Requires further checks on Mojave and Catalina with Xcode 11. Optionally
Homebrew with LLVM, Linux with Clang.